tests/wpt/tests/navigation-api/state/cross-document-getState-undefined.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    assert_equals(i.contentWindow.navigation.entries().length, 1);
    i.contentWindow.location.href = "?1";
    i.onload = t.step_func_done(() => {
      assert_equals(i.contentWindow.navigation.entries().length, 2);
      assert_equals(i.contentWindow.navigation.currentEntry.index, 1);
      assert_equals(i.contentWindow.navigation.entries()[0].getState(), undefined);
    });
  });
}, "Default behavior for entry.getState() for a non-current cross-document entry");
</script>
