LayoutTests/imported/w3c/web-platform-tests/navigation-api/state/cross-document-getState-undefined-expected.txt



FAIL Default behavior for entry.getState() for a non-current cross-document entry assert_equals: expected 2 but got 1
